+++ NancyFX/Program.cs
@@ -9,65 +9,43 @@ namespace NancyFX
 {
     class Program
     {
+        // Public static property to hold the Firebase client instance
         public static IFirebaseClient client { get; private set; }
 
-        private static Dictionary<string, string> LoadEnvFile(string filePath)
-        {
-            var envVariables = new Dictionary<string, string>();
-
-            foreach (var line in File.ReadAllLines(filePath))
-            {
-                var parts = line.Split('=', StringSplitOptions.RemoveEmptyEntries);
-
-                if (parts.Length != 2)
-                    continue;
-
-                envVariables[parts[0].Trim()] = parts[1].Trim();
-            }
-
-            return envVariables;
-        }
-
+        // Entry point of the program
         public static async Task Main(string[] args)
         {
-
-            var envVariables = LoadEnvFile(".env"); // Assuming .env file is in the root directory
-
-            // Extract necessary values
-            string authSecret = envVariables["AuthSecret"];
-            string basePath = envVariables["BasePath"];  // Replace with your actual environment variable key
-
+            // Configuration for Firebase connection
             IFirebaseConfig config = new FirebaseConfig
             {
-                // // Use the secret key from .env file
+                // AuthSecret and BasePath need to be provided for Firebase connection
                 // AuthSecret = authSecret,
                 // BasePath = basePath
-                // // Add other configuration details if necessary
-
-                AuthSecret = "mKrgFEbL8PmuavACKmhJ6q2XeQCSuqin5qGNKJMl",
-                BasePath = "https://taskterriers-39683-default-rtdb.firebaseio.com/"
             };
 
+            // Create a Firebase client instance with the configuration
             client = new FireSharp.FirebaseClient(config);
 
+            // Check if the client is connected to Firebase
             if (client != null)
             {
                 Console.WriteLine("Connected to Firebase");
             }
 
-            // Set up NancyFX host configuration
+            // Configure NancyFX hosting settings
             var hostConfigs = new HostConfiguration
             {
+                // Automatically create URL reservations to avoid permission issues
                 UrlReservations = new UrlReservations { CreateAutomatically = true }
             };
 
-            // Start the NancyFX server
-            using (var host = new NancyHost(new Uri("http://192.168.1.127:1234"), new CustomBootstrapper(), hostConfigs))
+            // Initialize and start the NancyFX server
+            using (var host = new NancyHost(new Uri("http://10.239.84.254:1234"), new CustomBootstrapper(), hostConfigs))
             {
                 host.Start();
-                Console.WriteLine("NancyFX is running on http://192.168.1.127:1234");
+                Console.WriteLine("NancyFX is running on http://10.239.84.254:1234");
 
-                // Perform Firebase operations
+                // Here you can perform Firebase operations such as retrieving users
                 // await GetUsers(client);
 
                 // Keep the server running indefinitely
